% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxB.R
\name{boxB}
\alias{boxB}
\title{paste the boxB() function from univOutl package}
\usage{
boxB(
  x,
  k = 1.5,
  method = "asymmetric",
  weights = NULL,
  id = NULL,
  exclude = NA,
  logt = FALSE
)
}
\description{
paste the boxB() function from univOutl package
}
