% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDC.R
\name{BWClusters}
\alias{BWClusters}
\title{Preliminary Calculation of Cluster vs Lineage Differentiation}
\usage{
BWClusters(dist_mat, c_cl)
}
\arguments{
\item{dist_mat}{A distance object that contains the distance between cells from dist() function}

\item{c_cl}{A vector of integers indicating the cluster to which each cell is allocated.}
}
\value{
A list about the results of Cluster vs Lineage Differentiation
\itemize{
  \item JaccardIndex - The Jaccard Index between cells
  \item signJaccardIndex - Edited Jaccard Index. The negative sign is added to show the cell is very close to other clusters.
  \item WithinDist - 	A distance list containing cell-to-cell distances within clusters
  \item BetweenDist - A distance list containing cell-to-cell distances between clusters
}
}
\description{
Preliminary Calculation of Cluster vs Lineage Differentiation
}
