% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDC.R
\name{BWClusters_Determination}
\alias{BWClusters_Determination}
\title{Precise Calculation of Cluster vs Lineage Differentiation}
\usage{
BWClusters_Determination(
  dist_mat,
  K,
  c_cl,
  cutoff = 0.3,
  checkcells = NULL,
  connectedCells = NULL,
  checksize = NULL
)
}
\arguments{
\item{dist_mat}{A distance object that contains the distance between cells from dist() function}

\item{K}{The number of clusters identified by preview step.}

\item{c_cl}{A vector of integers indicating the cluster to which each cell is allocated.}

\item{cutoff}{The density, determined using the Jaccard index cutoff, serves as a criterion for defining the connectivity of cells. The default value is 0.3.}

\item{checkcells}{The number of cells examined per cluster determines the identification of connected clusters. If left NULL, the default value is calculated as the maximum value of the minimum cluster size divided by 5 and 10.}

\item{connectedCells}{The minimum number of connected cells between clusters used to identify their connection. If left NULL, the default value is 10.}

\item{checksize}{The number of neighbors to consider for each check cell to determine their connectivity. If left NULL, the value is set to be equal to "checkcells".}
}
\value{
A list about the results of disconnected clusters detection
\itemize{
  \item DCcheck - the decision sentence
  \item ifConnected - a logical output. TRUE represents no disconnected clusters detected.
  \item Jaccardsummary - a table showing any two clusters are connected or not.
  \item clusterLocation - a table showing which two clusters are connected.
  \item K - The number of clusters identified.
  \item Clusters - A vector of integers indicating the cluster to which each cell is allocated.
  \item ConnectedCells - a list showing the connected cells between clusters.
  \item No_ConnectedCells - the number of the connected cells between clusters.
}
}
\description{
Precise Calculation of Cluster vs Lineage Differentiation
}
