% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_inshiny.R
\name{DE_edgeR}
\alias{DE_edgeR}
\title{Find top 30 DE genes between disconnected clusters}
\usage{
DE_edgeR(rawcounts, cls)
}
\arguments{
\item{rawcounts}{A raw count data matrix: row:genes, column:cells}

\item{cls}{Identified disconnected clusters}
}
\description{
Find top 30 DE genes between disconnected clusters
}
