% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_inshiny.R
\name{DE_seurat}
\alias{DE_seurat}
\title{Find DE genes between disconnected clusters}
\usage{
DE_seurat(rawcounts, cls)
}
\arguments{
\item{rawcounts}{A raw count data matrix: row:genes, column:cells}

\item{cls}{Identified disconnected clusters}
}
\description{
Find DE genes between disconnected clusters
}
