% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2D.R
\name{getDR_2D}
\alias{getDR_2D}
\title{Generate 2D Embeddings by different methods:}
\usage{
getDR_2D(norm_counts, method)
}
\arguments{
\item{norm_counts}{A normalized count data matrix: row:genes, column:cells}

\item{method}{choose one RD method: MDS, UMAP, PCA, TSNE, ICA}
}
\value{
A data frame containing coordinates of 2-dimensional embedding.
}
\description{
Generate 2D Embeddings by different methods:
}
