% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check2D.R
\name{GOFeval}
\alias{GOFeval}
\title{Goodness of Fit}
\usage{
GOFeval(dimred, alpha = NULL, fig = F)
}
\arguments{
\item{dimred}{A data frame. a 2D embedding from a DR method}

\item{alpha}{alpha value to calculate the alpha-convex hull. If left NULL, the alpha will be calculated based on the  first derivative.}

\item{fig}{A figure showing the cell distribution area.}
}
\value{
A list about the results of the GOF
\itemize{
  \item cellArea - The area of the alpha-convex hull of a sample of points.
  \item alpha - alpha value used to calculate the alpha-convex hull
  \item alphahull_obj - 	Object of class "ahull"
  \item occupiedRate - the proportion of alpha-convex hull area to the minimum circle that could enclose all cells in the embedding
}
}
\description{
Goodness of Fit
}
