% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.scLCA.R
\name{\%^\%}
\alias{\%^\%}
\title{Useful functions}
\usage{
M \%^\% power
}
\description{
This file contains a few useful functions for scLCA. This is from https://bitbucket.org/scLCA/single_cell_lca/src/master/
}
