% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDC.R
\name{HD_DCClusterscheck}
\alias{HD_DCClusterscheck}
\title{Differentiate Clusters vs Lineages in High-Dimensional Data}
\usage{
HD_DCClusterscheck(
  dist_mat,
  rawcounts,
  K = NULL,
  clust.max = 10,
  cutoff = 0.3,
  checkcells = NULL,
  connectedCells = NULL,
  checksize = NULL
)
}
\arguments{
\item{dist_mat}{A distance object that contains the distance between cells from dist() function}

\item{rawcounts}{A raw count data matrix: row:genes, column:cells}

\item{K}{The number of clusters if you have prior information. If left NULL, the number of clusters is estimated by scLCA.}

\item{clust.max}{The maximum number of clusters used in scLCA. The default number is 10.}

\item{cutoff}{The density, determined using the Jaccard index cutoff, serves as a criterion for defining the connectivity of cells. The default value is 0.3.}

\item{checkcells}{The number of cells examined per cluster determines the identification of connected clusters. If left NULL, the default value is calculated as the maximum value of the minimum cluster size divided by 5 and 10.}

\item{connectedCells}{The minimum number of connected cells between clusters used to identify their connection. If left NULL, the default value is 10.}

\item{checksize}{The number of neighbors to consider for each check cell to determine their connectivity. If left NULL, the value is set to be equal to "checkcells".}

\item{seed}{The pseudorandom number used to reproduce the output.}
}
\value{
A list about the results of disconnected clusters detection.
}
\description{
Differentiate Clusters vs Lineages in High-Dimensional Data
}
