% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_inshiny.R
\name{HVGs_GO}
\alias{HVGs_GO}
\title{GSEA genes between disconnected clusters}
\usage{
HVGs_GO(normcounts, OrgDb = "org.Hs.eg.db", geneVar = NULL)
}
\arguments{
\item{normcounts}{A normalized count data matrix: row:genes, column:cells}

\item{OrgDb}{OrgDb}
}
\description{
GSEA genes between disconnected clusters
}
