% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_inshiny.R
\name{HVGs_quick}
\alias{HVGs_quick}
\title{Find HVGs by using quick scran}
\usage{
HVGs_quick(norm_counts)
}
\arguments{
\item{norm_counts}{A normalized count data matrix: row:genes, column:cells}
}
\description{
Find HVGs by using quick scran
}
