% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_inshiny.R
\name{HVGs_scran}
\alias{HVGs_scran}
\title{Find HVGs by using scran package}
\usage{
HVGs_scran(norm_counts)
}
\arguments{
\item{norm_counts}{A normalized count data matrix: row:genes, column:cells}
}
\description{
Find HVGs by using scran package
}
