% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDC.R
\name{JaccardIndex_fun}
\alias{JaccardIndex_fun}
\title{Jaccard Index of cells}
\usage{
JaccardIndex_fun(w_vec, b_vec)
}
\arguments{
\item{w_vec}{A distance vector containing cell-to-cell distances within clusters}

\item{b_vec}{A distance vector containing cell-to-cell distances between clusters}
}
\value{
A list containing the Jaccard Index between cells
\itemize{
  \item JaccardIndex - The Jaccard Index between cells
  \item signJaccardIndex - Edited Jaccard Index. The negative sign is added to show the cell is very close to other clusters.
}
}
\description{
Jaccard Index of cells
}
