% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDC.R
\name{LD_DCClusterscheck}
\alias{LD_DCClusterscheck}
\title{Differentiate Clusters vs Lineages in Embedding}
\usage{
LD_DCClusterscheck(
  embedding,
  cutoff = 0.1,
  max.nc = 5,
  K = NULL,
  checkcells = NULL,
  connectedCells = 1,
  checksize = NULL
)
}
\arguments{
\item{embedding}{A data frame of the embedding.}

\item{cutoff}{The density, determined using the Jaccard index cutoff, serves as a criterion for defining the connectivity of cells. The default value is 0.1.}

\item{max.nc}{The maximum number of clusters used in NbClust. The default number is 5.}

\item{K}{The number of clusters if you have prior information. If left NULL, the number of clusters is estimated by NbClust.}

\item{checkcells}{The number of cells examined per cluster determines the identification of connected clusters. If left NULL, the default value is calculated as the maximum value of the minimum cluster size divided by 5 and 10.}

\item{connectedCells}{The minimum number of connected cells between clusters used to identify their connection. The default value is 1. If left NULL, the default value is 10.}

\item{checksize}{The number of neighbors to consider for each check cell to determine their connectivity. If left NULL, the value is set to be equal to "checkcells".}
}
\value{
A list about the results of disconnected clusters detection.
}
\description{
Differentiate Clusters vs Lineages in Embedding
}
