% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickHVG.R
\name{quick_model_gene_var}
\alias{quick_model_gene_var}
\title{Quick HVG (simplified from scran)}
\usage{
quick_model_gene_var(normcounts)
}
\arguments{
\item{normcounts}{A normalized count data matrix: row:genes, column:cells}
}
\description{
Quick HVG (simplified from scran)
}
