% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_cal.R
\name{score_cal}
\alias{score_cal}
\title{Scoring System}
\usage{
score_cal(scoredf, SimiRetaincutoff = 0, GOFcutoff = 0.53, URATEcutoff = 0.029)
}
\arguments{
\item{scoredf}{The data frame of containing all assessment content}

\item{SimiRetaincutoff}{The cutoff of the structure similarity. The default value is 0.8.}
}
\value{
a dataframe with scores for each embedding.
}
\description{
Scoring System
}
