% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepTraj.R
\name{segFormat}
\alias{segFormat}
\title{Format line segments (when using Slingshot)}
\usage{
segFormat(fittedLines)
}
\arguments{
\item{fittedLines}{principle curves obtained form the slingCurves() function in Slingshot}
}
\description{
Format line segments (when using Slingshot)
}
