% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check2D.R
\name{Similaritycheck}
\alias{Similaritycheck}
\title{Structure Similarity between Orignal data and Embedding}
\usage{
Similaritycheck(clusters, dimred, fig = F)
}
\arguments{
\item{clusters}{object from \code{DCClusterscheck} fucntion.}

\item{dimred}{A data frame. a 2D embedding from a DR method and the row name is the cell name.}

\item{fig}{A figure showing the clusters identified in raw data in the embedding}

\item{normcounts}{a normalized count data matrix: row:genes, column:cells}
}
\value{
A list about the results of structure similarity check between high- and low- dimenaional data.
\itemize{
  \item SimilarityLevel - Same group level
  \item GoodRate - High same group level percentage
  \item all - Same group level for each cell
  \item cls - clusters
}
}
\description{
Structure Similarity between Orignal data and Embedding
}
