% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testHomo.R
\name{step1_testHomogeneous}
\alias{step1_testHomogeneous}
\title{Test Homogeneous}
\usage{
step1_testHomogeneous(norm_counts, HVGs = NULL, g = 100, num.sim = 20000)
}
\arguments{
\item{norm_counts}{A normalized count data matrix: row:genes, column:cells}

\item{HVGs}{Highly variable genes. If there is no input of the highly variable genes, they will be calcualted by `modelGeneVar()` function.}

\item{g}{The number of highly variable genes to test}

\item{num.sim}{The number of simulations generated for permutation test on spearman correlation. Default is 20000. Reduce this to improve speed.}
}
\value{
A list about the results of trajectory signal detection
\itemize{
  \item signal_pct - the percentage of trajectory signal detected. If the percentage exceeds 50%, we conclude that the trajectory signal detected.
  \item decision - the decision sentence.
}
}
\description{
Test Homogeneous
}
