% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testHomo.R
\name{testHomogeneous}
\alias{testHomogeneous}
\title{Test Homogeneous}
\usage{
testHomogeneous(HVGs, norm_counts, n = 100, num.sim = 20000)
}
\arguments{
\item{HVGs}{Highly variable genes}

\item{norm_counts}{A normalized count data matrix: row:genes, column:cells}

\item{n}{The number of highly variable genes to test}

\item{num.sim}{The number of simulations generated for permutation test on spearman correlation}
}
\value{
A list about the results of trajectory signal detection
\itemize{
  \item signal_pct - the percentage of trajectory signal detected. If the percentage exceeds 50%, we conclude that the trajectory signal detected.
  \item decision - the decision sentence.
}
}
\description{
Test Homogeneous
}
