% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepTraj.R
\name{prepTraj}
\alias{prepTraj}
\title{Trajectory Evaluation Object}
\usage{
prepTraj(dimred, PT, fitLine)
}
\arguments{
\item{dimred}{A data frame. a 2D embedding from a DR method and the row name is the cell name.}

\item{PT}{A data frame or a vector. estimated pseudotime from TI. Each column contains the PT for one lineage.}

\item{fitLine}{the fitted curve line got from TI containing line segments between pairs of points.}
}
\value{
a list containing related information used for future analysis.
\itemize{
  \item Embedding - A data frame. a 2D embedding from a DR method
  \item pse - A data frame. a scaled pseudotime
  \item fitLine - A data frame. line segments between pairs of points.
}
}
\description{
Trajectory Evaluation Object
}
