% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAmbigcells.R
\name{UshapeDetector}
\alias{UshapeDetector}
\title{Check the shape of trajectory curves}
\usage{
UshapeDetector(obj, outlierdetect = "neutral", fig = F)
}
\arguments{
\item{obj}{A basic trajectory evaluation object from \code{prepTraj} function.}

\item{outlierdetect}{The method used to detect the outlier: "resistant", "adjbox", "asymmetric", "neutral". The default method is "neutral".}

\item{fig}{A figure showing the ambiguous cells within the embedding.}
}
\value{
A list about the results of structure similarity check between high- and low- dimenaional data.
\itemize{
  \item AmbiguousCells - Vector of ambiguous cells (those considered hard to estimate accurately).
  \item NoACells - Number of ambiguous cells.
  \item Ambpct - Percentage of ambiguous cells.
}
}
\description{
Check the shape of trajectory curves
}
